Refactoring the file structure, closes #14
